requirements.txt
matrix-nio[e2e]
PyYAML
requests
